+++ Mnema.Providers/Managers/Publication/MangaPublicationExtensions.cs
@@ -2,7 +2,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Mnema.Providers;
+namespace Mnema.Providers.Managers.Publication;
 
 internal interface IPreDownloadHook
 {
