+++ Mnema.Providers/Managers/Publication/Publication.Loading.cs
@@ -11,7 +11,7 @@ using Mnema.Models.DTOs.Content;
 using Mnema.Models.Entities.Content;
 using Mnema.Models.Publication;
 
-namespace Mnema.Providers;
+namespace Mnema.Providers.Managers.Publication;
 
 internal partial class Publication
 {
