+++ Mnema.Providers/Managers/QBit/QBitContentManager.Downloading.cs
@@ -13,7 +13,7 @@ using Mnema.Models.DTOs.Content;
 using Mnema.Models.Entities.Content;
 using QBittorrent.Client;
 
-namespace Mnema.Providers.QBit;
+namespace Mnema.Providers.Managers.QBit;
 
 internal partial class QBitContentManager
 {
